Move unification stuff to its own module
