+++ covenant.cabal
@@ -106,6 +106,7 @@ library
     Covenant.Internal.Rename
     Covenant.Internal.TyExpr
     Covenant.Internal.Type
+    Covenant.Internal.Unification
 
   build-depends:
     QuickCheck ==2.15.0.1,
